Move template components with their stories into seperate directories
