+++ .vscode/settings.json
@@ -31,10 +31,14 @@
         "03_organisms": "Components",
         "Footer": "Components",
         "Header": "Components",
-		"04_templates": "Layout",
+        "04_templates": "Layout",
+        "MainLayout": "Layout",
+		"PageTemplate": "Layout",
+		"PostTemplate": "Layout",
 		"05_pages": "Log",
 		"06_posts": "Docs",
-		"07_images": "Images",
+        "07_images": "Images",
+        "icon_logos": "Images",
 		"08_icons": "Svg",
 		"09_utilities": "Utils",
 		"10_styles": "Sass",
