+++ src/pages/404.js
@@ -8,7 +8,7 @@
 
 import React from "react"
 import { Link } from "gatsby"
-import Layout from "../04_templates/MainLayout"
+import Layout from "../04_templates/MainLayout/MainLayout"
 
 const NotFoundPage = () => (
     <Layout>
